LRequest: use pidfd to avoid PID recycling races
If the client process exits (but hands over the socket to another
process), its PID may get recycled eventually.  If the new process
with the same PID belongs to a different cgroup, the Lua script may
grant the wrong credentials.  This vulnerability is fixed by using a
pidfd (obtained via `SO_PEERPIDFD`) which is compared with `ucred.pid`
after reading `/proc/PID/cgroup`.